// models/pokemon.ts
export interface RandomResponse {
  id: number;
  image: string;
  names: string[];
}

export interface VerifyResponse {
  correct: boolean;
  name: string;
}